🚚 move recovered artifact tests into tests directories
